refactor: corrected code by comments
